Implement command classes (CreateDeadlineForStudentCommand, CreateStudentCommand and CreateToDoForStudentCommand) and their respective test cases
Code that is depended on other methods(that are not yet implemented) are
commented out.

Other methods such as equals method in AbsolutePath class and duplicate
method in RootOperation class and GroupOperation class.

Test cases for CreateStudentCommand has been commented out due to errors.
Theequals and toString methods of CreateDeadlineForStudentCommand and
CreateToDoForStudentCommand have test cases.